+++ tutor/commands/context.py
@@ -1,0 +1,4 @@
+# pylint: disable=too-few-public-methods
+class Context:
+    def __init__(self, root):
+        self.root = root